support for tugboat messages
when translating a classic query, tugboat may add error and warning messages. these changes support bumblebee display of the mesages.
this is a partial solution, it does not include display of tugboat messages for objects searches.
this commit includes the qs function from utils.js because it is not in this branch.  at some point the qs function from utils should be used.