Fix history loop issue with /search
When an error occurs on the route /search, it defaults back to index
page.  But this error will happen with any bare (no params) search, so
you can get into a loop where the back button is constantly re-pushing
the index-page.
This now replaces the history entry on error, which should fix it.