Simplify formatDate function to fix date parsing
JS built-in date seems to have issues when the day is 00, causing it to
default to the last day of the previous month.

The fix was to rewrite using regex to parse the incoming date, then use
momentjs to parse and output.

Before it was using a combination of Date and jquery.datepicker