Fix/Update test
Removing the analytics test for now.  This test was mainly used to make
sure the array was properly pushed to, and that it does the right thing
if the global is undefined.  I think this module is simple enough to
leave as a proxy to the GTM call.  We can add more testing for the
logger in the future.