Attempt to pass a good user_id to GA
Decided to just watch the USER_ANNOUNCEMENT event since this should be called by the pubsub after the user is updated.

This value must be hashed though to avoid leaking user information, so it may limit to browsers supporting `window.crypto`.

@see https://caniuse.com/cryptography

Seems like good coverage as of now, so hopefully safe to use this here without polyfills.