Fix module timeout issue
Even though the configured bundles have a timeout set of 30 seconds, and
the shim sets it to `0` the initial modules are subject to a `7` second
timeout (default).

This resets all of the timeouts, including the default to be 30 seconds.
This should allow even slow connections to wait for pending requests.