Add filter for incoming queries that do not have `fl`
The way it works is that the search cycle is initiated during the navigation transition.  This query is then passed to any subscribers, but also gets processed even though the results are not used.  This attempts to filter out those types of queries from getting processed, but does allow pub/sub events so that derived requests can be created.